Bug 1168856 - Part 7: split out Downloader.swift to make logging make sense, and let Xcode have its way with the project file.
